requirements.txt
scipy==1.4.1
tqdm==4.61.2
numpy==1.18.5
h5py==2.10.0
torch==1.9.0
opencv_python==4.5.1.48
torchvision==0.10.0
albumentations==1.1.0
matplotlib==3.4.3
pandas==1.3.3
Pillow==8.3.2
pytorch_lightning==1.4.9
scikit-image==0.18.3
torchinfo==1.5.3
